add console package
